styling plant card modal
