+++ src/components/PlantCardModal/PlantCardModal.scss
@@ -1,0 +1,71 @@
+@import '../../_utils';
+
+.modalWrapper {
+    padding: rem(20px);
+    max-width: rem(800px);
+    margin-left: rem(10px);
+    margin-right: rem(10px);
+    border-radius: 10px;
+}
+ 
+.modalImg {
+    width: 70%;
+    height: auto;
+    object-fit: cover;
+    border-radius: rem(10px);
+    box-shadow: 0 rem(4px) rem(8px) $gray-10;
+    margin-bottom: rem(30px);
+    margin-top: rem(30px);
+}
+
+.modalName {
+    font-family: $mono;
+    font-size: rem(20px);
+}
+
+.modalLatinName {
+    font-family: $mono;
+    font-style: italic;
+    font-size: rem(16px);
+    font-weight: 400;
+    margin-bottom: rem(20px);
+    color: $gray;
+}
+
+.modalDescription {
+    max-width: rem(800px);
+    margin-bottom: rem(30px);
+    text-align: left;
+}
+
+.modalSunWaterWrapper {
+    width: 100%;
+    max-width: rem(800px);
+    margin-bottom: rem(10px);
+    text-align: left;
+    
+    svg {
+        margin-right: rem(15px);
+        font-size: 18px;
+    }
+}    
+
+.modalShareButton {
+    margin-top: rem(20px);
+    margin-bottom: rem(20px);
+    padding: rem(5px) rem(25px);
+    background-color: $white;
+    border: 2px solid $black;
+    color: $black;
+    border-radius: 10px;
+
+    button {
+        font-weight: 600;
+    }
+}
+
+.modalShareButton:hover {
+    background-color: $green-light;
+    border: 2px solid $green-light;
+    color: $white;
+}
